Abortive attempt at using spec to parse IR statements.
Once again, it's simpler to be more verbose and not worry about parsing at
all. I think spec would be useful as a parser if we were forced to consume a
data DSL we did not control, but that's not my issue here.

If I start parsing IR text, spec could be a great way to turn the output of the
lexer into data structures. But that's a problem for the future.