Start of a rewrite.
It's not even working yet but it's obviously off to a bad start.

Instructions themselves should be more verbose; don't try and emulate the syntax
of the textual IR, that's a fool's errand.

Also don't generate strings directly. Generate an intermediate
representation (maybe just a tree of tokens), and generate text from that.

How many intermediate representations is that now? 5?