Back to direct string generation.
I kept the more verbose instruction descriptions, but the formatting of the code
is too tightly coupled to the semantics of the code, so we would need two passes
to generate the token stream and then format it.

To save time, I'm just generating the tokens and the formatting in one pass.

It's fast enough for present purposes, and far more important, it works!