Attempt the nth.
Fully separating signal transformation from aggregation might be impossible. At
the very least it's starting to appear undesirable.

In order for a transducer to discard a message, it needs to have a way to
pass. In order to return multiple messages it needs to be able to call further
down the chain multiple times.

Hickey found a very good spot in the design space where everything
composes. Well mostly. To have prefixes, you need to use state, and sufficies
require an odd calling convention. There still feels like room for improvement,
especially regarding state, but I haven't found it.