First class bitstrings are going to simplify the lower level.
High level operations on machine code. I like that. Being able to destructure
structs that don't correspond to anything compiled in our codebase will be
important for interop. How else are you going to read a specification at runtime
and figure out how to read an incoming binary data stream on the fly?