Incomplete attempt to replace interpreter channels with streams.
There's still something off about the way I'm looking at this.

There's a dichotomy where some things return streams and others return
values.

The solution I think I want is to make everything take and return streams.

But then primitives need to be special because they need values. And the root
needs to be special because the repl needs to print things. And if that many
things are special, nothing is special anymore.