More notes.
I'm getting tied up with mapping between lisp and message passing.

Do accumulators live in the environment? But then the environment is mutable, if
only in that the "current state" is always changing. But is that current state
part of the environment or the runtime? Would that mean the runtime environment
is a separate thing from the lexical environment of the message currently being
sent? Do programmed constructs need access to that environment? Would it be safe
to provide?

Well questions are good, but I need some answers too one day.